Update docs.md
Throughout the api guide there are issues with linking related to methods calls.  I believe this is due to the markdown link syntax of []() being disturbed by the use of the () after the method name, i.e. 'SoundCache.getSound()'.  I think this can be fixed with escaping the method's argument parenthesis.  Hard to tell if it fixes it from editing on github as it already accounts for that issue, but I will fix the other ones I see just in case as I'm pretty certain that is the issue but could always be wrong.