CLEAN UP CODING STANDARDS
I noticed that this list had references to numbered sections, but the numbers were missing except for the final section. 

- Added numbers to all the sections to match as close to the github wiki coding standards
- Cleaned up for general stylistic readability
- Fixed the links referencing other sections
- Cleared up some of the examples given

I apologize in advance if there was a plan to remove the number system completely.  If so, please ignore these changes.

Final note :
Please test the reference links as they may need to be compiled first from post merging before applicable.